requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
coverage==7.3.2
    # via -r requirements.in
crispy-bootstrap5==2023.10
    # via -r requirements.in
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
django==4.2.7
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-admin-autocomplete-filter
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-import-export
    #   django-recaptcha
    #   django-reversion
    #   model-bakery
django-admin-autocomplete-filter==0.7.1
    # via -r requirements.in
django-admin-rangefilter==0.11.2
    # via -r requirements.in
django-crispy-forms==2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-grappelli==3.0.8
    # via -r requirements.in
django-import-export==3.3.3
    # via -r requirements.in
django-recaptcha==4.0.0
    # via -r requirements.in
django-reversion==5.0.8
    # via -r requirements.in
django-simple-history==3.4.0
    # via -r requirements.in
django-watson==1.6.3
    # via -r requirements.in
django-widget-tweaks==1.5.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
factory-boy==3.3.0
    # via -r requirements.in
faker==12.0.1
    # via
    #   factory-boy
    #   mixer
iniconfig==2.0.0
    # via pytest
markuppy==1.14
    # via tablib
mixer==7.2.2
    # via -r requirements.in
model-bakery==1.17.0
    # via -r requirements.in
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==23.2
    # via
    #   pytest
    #   pytest-sugar
pluggy==1.3.0
    # via pytest
psycopg[binary]==3.1.13
    # via -r requirements.in
psycopg-binary==3.1.13
    # via psycopg
pytest==7.4.3
    # via
    #   pytest-django
    #   pytest-sugar
pytest-django==4.7.0
    # via -r requirements.in
pytest-sugar==0.9.7
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
pyyaml==6.0.1
    # via tablib
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via
    #   django-import-export
    #   tablib
termcolor==2.3.0
    # via pytest-sugar
typing-extensions==4.8.0
    # via psycopg
whitenoise==6.6.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
